Requested changes: community referenced by module id. Only one rating possible per user and per community. Add flag to community whether logged in user has joined it.

+++ src/main/java/com/agileavengers/icuconnectbackend/model/Rating.java
@@ -24,9 +24,9 @@ public class Rating {
 
     @ManyToOne
     private Community community;
-    private Integer teaching;
-    private Integer content;
-    private Integer workload;
+    private Double teaching;
+    private Double content;
+    private Double workload;
 
     private String text;
